Initial commit of the new 3.0 website
In the first iteration of the website we had plans to build an ecosystem
around it with custom themes and plugins. This never came to fruition
but the unnecessary infra was still there.

For phpDocumentor 3.0 we needed a new design, but because the codebase
was out of date it was easier to reboot than to try to get the old site
working again.

This is the initial work to make the website available by using Gatsby
to generate the site based on React Components.